Improve: Add messageBox for deleting folder in grid view

+++ bundles/AdminBundle/public/js/pimcore/object/folder.js
@@ -425,11 +425,15 @@ pimcore.object.folder = Class.create(pimcore.object.abstract, {
 
 
     remove: function () {
-        var options = {
+        const options = {
             "elementType" : "object",
             "id": this.id
         };
-        pimcore.elementservice.deleteElement(options);
+        Ext.MessageBox.confirm(t('delete_folder'), `${t('delete_entire_folder_question')} ${this.data.general.o_key}?`, function(btn){
+            if (btn === 'yes') {
+                pimcore.elementservice.deleteElement(options);
+            }
+        });
     },
 
     isAllowed : function (key) {
